code optimize & line comments add
