+++ lib/core/constants/colors_constants.dart
@@ -5,3 +5,4 @@ const Color kGreen = Color(0xFF76B536);
 const Color kLightGreenAccent = Color(0xFFEBFDF2);
 const Color kTextFieldFilledColor = Color.fromARGB(255, 220, 232, 225);
 const Color kBlack = Color(0xFF000000);
+const Color kGrey = Color(0xFF808080);
