+++ lib/core/constants/text_style.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:w3/core/constants/colors_constants.dart';
+TextStyle kH1TextStyleLogInScreen=  TextStyle(
+    fontFamily: 'Poppins',
+    fontSize: 30.0.sp,
+    fontWeight: FontWeight.bold);
+
+TextStyle kH2TextStyleLogInScreen=  TextStyle(
+  fontFamily: 'Poppins',
+  fontSize: 14.0.sp,
+  color: kGrey,
+  fontWeight: FontWeight.w300,);
+
+ TextStyle k4TextStyleLogInScreen=  TextStyle(
+  fontFamily: 'Poppins',
+  fontSize: 10.0.sp,
+  color: kGrey,
+ );
+
+TextStyle kLongInButtonTextStyle=TextStyle(fontFamily: "Poppins",
+fontSize: 15.sp);
+
+TextStyle kH3TextStyleLogInScreen=  TextStyle(
+  fontFamily: 'Poppins',
+  fontSize: 14.0.sp,
+  color: kGrey,
+);
+TextStyle kH3TextStyleSignUpTextInLogInScreen=TextStyle(
+decoration: TextDecoration.underline,
+decorationThickness: 2.h,
+fontSize: 12.sp,
+fontWeight: FontWeight.bold,
+);
